+++ src/distilabel/steps/clustering/text_clustering.py
@@ -74,7 +74,7 @@ class TextClustering(TextClassification, GlobalTask):
         Generate labels for a set of texts using clustering:
 
         ```python
-        from distilabel.llms import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
         from distilabel.steps import UMAP, DBSCAN, TextClustering
         from distilabel.pipeline import Pipeline
