+++ src/distilabel/steps/embeddings/nearest_neighbour.py
@@ -84,7 +84,7 @@ class FaissNearestNeighbour(GlobalStep):
         Generating embeddings and getting the nearest neighbours:
 
         ```python
-        from distilabel.embeddings.sentence_transformers import SentenceTransformerEmbeddings
+        from distilabel.models import SentenceTransformerEmbeddings
         from distilabel.pipeline import Pipeline
         from distilabel.steps import EmbeddingGeneration, FaissNearestNeighbour, LoadDataFromHub
