+++ src/distilabel/steps/tasks/apigen/generator.py
@@ -88,7 +88,7 @@ class APIGenGenerator(Task):
 
         ```python
         from distilabel.steps.tasks import ApiGenGenerator
-        from distilabel.llms import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         llm=InferenceEndpointsLLM(
             model_id="meta-llama/Meta-Llama-3.1-70B-Instruct",
@@ -138,7 +138,7 @@ class APIGenGenerator(Task):
 
         ```python
         from distilabel.steps.tasks import ApiGenGenerator
-        from distilabel.llms import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         llm=InferenceEndpointsLLM(
             model_id="meta-llama/Meta-Llama-3.1-70B-Instruct",
