+++ src/distilabel/steps/tasks/complexity_scorer.py
@@ -67,7 +67,7 @@ class ComplexityScorer(Task):
 
         ```python
         from distilabel.steps.tasks import ComplexityScorer
-        from distilabel.llms.huggingface import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         # Consider this as a placeholder for your actual LLM.
         scorer = ComplexityScorer(
@@ -91,7 +91,7 @@ class ComplexityScorer(Task):
 
         ```python
         from distilabel.steps.tasks import ComplexityScorer
-        from distilabel.llms.huggingface import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         # Consider this as a placeholder for your actual LLM.
         scorer = ComplexityScorer(
