+++ src/distilabel/steps/tasks/evol_instruct/base.py
@@ -75,7 +75,7 @@ class EvolInstruct(Task):
 
         ```python
         from distilabel.steps.tasks import EvolInstruct
-        from distilabel.llms.huggingface import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         # Consider this as a placeholder for your actual LLM.
         evol_instruct = EvolInstruct(
@@ -96,7 +96,7 @@ class EvolInstruct(Task):
 
         ```python
         from distilabel.steps.tasks import EvolInstruct
-        from distilabel.llms.huggingface import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         # Consider this as a placeholder for your actual LLM.
         evol_instruct = EvolInstruct(
@@ -124,7 +124,7 @@ class EvolInstruct(Task):
 
         ```python
         from distilabel.steps.tasks import EvolInstruct
-        from distilabel.llms.huggingface import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         # Consider this as a placeholder for your actual LLM.
         evol_instruct = EvolInstruct(
