+++ src/distilabel/steps/tasks/evol_instruct/evol_complexity/base.py
@@ -67,7 +67,7 @@ class EvolComplexity(EvolInstruct):
 
         ```python
         from distilabel.steps.tasks import EvolComplexity
-        from distilabel.llms.huggingface import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         # Consider this as a placeholder for your actual LLM.
         evol_complexity = EvolComplexity(
