+++ src/distilabel/steps/tasks/evol_instruct/evol_complexity/generator.py
@@ -65,7 +65,7 @@ class EvolComplexityGenerator(EvolInstructGenerator):
 
         ```python
         from distilabel.steps.tasks import EvolComplexityGenerator
-        from distilabel.llms.huggingface import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         # Consider this as a placeholder for your actual LLM.
         evol_complexity_generator = EvolComplexityGenerator(
