+++ src/distilabel/steps/tasks/evol_instruct/generator.py
@@ -81,7 +81,7 @@ class EvolInstructGenerator(GeneratorTask):
 
         ```python
         from distilabel.steps.tasks import EvolInstructGenerator
-        from distilabel.llms.huggingface import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         # Consider this as a placeholder for your actual LLM.
         evol_instruct_generator = EvolInstructGenerator(
