+++ src/distilabel/steps/tasks/evol_quality/base.py
@@ -71,7 +71,7 @@ class EvolQuality(Task):
 
         ```python
         from distilabel.steps.tasks import EvolQuality
-        from distilabel.llms.huggingface import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         # Consider this as a placeholder for your actual LLM.
         evol_quality = EvolQuality(
