+++ src/distilabel/steps/tasks/genstruct.py
@@ -73,7 +73,7 @@ class Genstruct(Task):
 
         ```python
         from distilabel.steps.tasks import Genstruct
-        from distilabel.llms.huggingface import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         # Consider this as a placeholder for your actual LLM.
         genstruct = Genstruct(
