+++ src/distilabel/steps/tasks/prometheus_eval.py
@@ -138,7 +138,7 @@ class PrometheusEval(Task):
 
         ```python
         from distilabel.steps.tasks import PrometheusEval
-        from distilabel.llms import vLLM
+        from distilabel.models import vLLM
 
         # Consider this as a placeholder for your actual LLM.
         prometheus = PrometheusEval(
@@ -175,7 +175,7 @@ class PrometheusEval(Task):
 
         ```python
         from distilabel.steps.tasks import PrometheusEval
-        from distilabel.llms import vLLM
+        from distilabel.models import vLLM
 
         # Consider this as a placeholder for your actual LLM.
         prometheus = PrometheusEval(
@@ -212,7 +212,7 @@ class PrometheusEval(Task):
 
         ```python
         from distilabel.steps.tasks import PrometheusEval
-        from distilabel.llms import vLLM
+        from distilabel.models import vLLM
 
         # Consider this as a placeholder for your actual LLM.
         prometheus = PrometheusEval(
@@ -252,7 +252,7 @@ class PrometheusEval(Task):
 
         ```python
         from distilabel.steps.tasks import PrometheusEval
-        from distilabel.llms import vLLM
+        from distilabel.models import vLLM
 
         # Consider this as a placeholder for your actual LLM.
         prometheus = PrometheusEval(
