+++ src/distilabel/steps/tasks/quality_scorer.py
@@ -67,7 +67,7 @@ class QualityScorer(Task):
 
         ```python
         from distilabel.steps.tasks import QualityScorer
-        from distilabel.llms.huggingface import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         # Consider this as a placeholder for your actual LLM.
         scorer = QualityScorer(
@@ -102,7 +102,7 @@ class QualityScorer(Task):
 
         ```python
         from distilabel.steps.tasks import QualityScorer
-        from distilabel.llms.huggingface import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         scorer = QualityScorer(
             llm=InferenceEndpointsLLM(
