+++ src/distilabel/steps/tasks/self_instruct.py
@@ -66,7 +66,7 @@ class SelfInstruct(Task):
 
         ```python
         from distilabel.steps.tasks import SelfInstruct
-        from distilabel.llms.huggingface import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         self_instruct = SelfInstruct(
             llm=InferenceEndpointsLLM(
