+++ src/distilabel/steps/tasks/sentence_transformers.py
@@ -108,7 +108,7 @@ class GenerateSentencePair(Task):
 
         ```python
         from distilabel.steps.tasks import GenerateSentencePair
-        from distilabel.llms import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         generate_sentence_pair = GenerateSentencePair(
             triplet=True, # `False` to generate only positive
@@ -128,7 +128,7 @@ class GenerateSentencePair(Task):
         Generating semantically similar sentences:
 
         ```python
-        from distilabel.llms import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
         from distilabel.steps.tasks import GenerateSentencePair
 
         generate_sentence_pair = GenerateSentencePair(
@@ -150,7 +150,7 @@ class GenerateSentencePair(Task):
 
         ```python
         from distilabel.steps.tasks import GenerateSentencePair
-        from distilabel.llms import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         generate_sentence_pair = GenerateSentencePair(
             triplet=True, # `False` to generate only positive
@@ -171,7 +171,7 @@ class GenerateSentencePair(Task):
 
         ```python
         from distilabel.steps.tasks import GenerateSentencePair
-        from distilabel.llms import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         generate_sentence_pair = GenerateSentencePair(
             triplet=True, # `False` to generate only positive
@@ -192,7 +192,7 @@ class GenerateSentencePair(Task):
 
         ```python
         from distilabel.steps.tasks import GenerateSentencePair
-        from distilabel.llms import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         generate_sentence_pair = GenerateSentencePair(
             triplet=True, # `False` to generate only positive
@@ -214,7 +214,7 @@ class GenerateSentencePair(Task):
 
         ```python
         from distilabel.steps.tasks import GenerateSentencePair
-        from distilabel.llms import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         generate_sentence_pair = GenerateSentencePair(
             triplet=True, # `False` to generate only positive
@@ -237,7 +237,7 @@ class GenerateSentencePair(Task):
 
         ```python
         from distilabel.steps.tasks import GenerateSentencePair
-        from distilabel.llms import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         generate_sentence_pair = GenerateSentencePair(
             triplet=True, # `False` to generate only positive
