+++ src/distilabel/steps/tasks/structured_generation.py
@@ -52,7 +52,7 @@ class StructuredGeneration(Task):
 
         ```python
         from distilabel.steps.tasks import StructuredGeneration
-        from distilabel.llms import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         structured_gen = StructuredGeneration(
             llm=InferenceEndpointsLLM(
@@ -109,7 +109,7 @@ class StructuredGeneration(Task):
 
         ```python
         from distilabel.steps.tasks import StructuredGeneration
-        from distilabel.llms import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         structured_gen = StructuredGeneration(
             llm=InferenceEndpointsLLM(
