+++ src/distilabel/steps/tasks/text_classification.py
@@ -90,7 +90,7 @@ class TextClassification(Task):
 
         ```python
         from distilabel.steps.tasks import TextClassification
-        from distilabel.llms.huggingface import InferenceEndpointsLLM
+        from distilabel.models import InferenceEndpointsLLM
 
         llm = InferenceEndpointsLLM(
             model_id="meta-llama/Meta-Llama-3.1-70B-Instruct",
